int->float

+++ target_equal.py
@@ -1,9 +1,9 @@
 print("計算普通相等吞吐量")
 print("========================")
 s = input('請輸入單個host的吞吐量 (s1,s2,s3...)：')   
-s = list(map(int, s.split(',')))
+s = list(map(float, s.split(',')))
 c = input('請輸入多個host並行的吞吐量 (c1,c2,c3...)：')   
-c = list(map(int, c.split(',')))
+c = list(map(float, c.split(',')))
 
 oneDsi_sum = 0
 for ss in s:
